Backend/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.2
urllib3==1.26.9
